requirements.txt
aiogram==3.13.1
gspread==6.1.2
google-auth==2.35.0
aiosqlite==0.20.0
python-dotenv==1.0.1
